+++ CHANGES.rst
@@ -83,6 +83,9 @@ New Features
 
 - ``astropy.nddata``
 
+    - Added ``subsampling`` to extract_array to allow for the extraction of
+    a subsampled patch centered on a set of coordinates.
+
 - ``astropy.stats``
 
   - Added ``biweight_midcovariance`` function. [#5777]